refactor: 영화 목록 화면 코드 리팩터링
