+++ app/src/main/java/woowacourse/movie/view/movie/MovieListContract.kt
@@ -1,10 +1,10 @@
 package woowacourse.movie.view.movie
 
-import woowacourse.movie.view.movie.model.MovieUiModel
+import woowacourse.movie.view.movie.adapter.MovieListItem
 
 interface MovieListContract {
     interface View {
-        fun showMovieList(movies: List<MovieUiModel>)
+        fun showMovieList(movies: List<MovieListItem>)
     }
 
     interface Presenter {
