+++ app/src/main/java/woowacourse/movie/view/movie/MovieListPresenter.kt
@@ -1,6 +1,9 @@
 package woowacourse.movie.view.movie
 
 import woowacourse.movie.common.DummyData
+import woowacourse.movie.view.movie.adapter.MovieListItem
+import woowacourse.movie.view.movie.model.AdUiModel
+import woowacourse.movie.view.movie.model.MovieUiModel
 import woowacourse.movie.view.movie.model.toUiModel
 
 class MovieListPresenter(
@@ -8,6 +11,22 @@ class MovieListPresenter(
 ) : MovieListContract.Presenter {
     override fun loadMovieListScreen() {
         val movies = DummyData.movies.map { it.key.toUiModel() }
-        view.showMovieList(movies)
+        val listWithAds = buildListWithAds(movies)
+        view.showMovieList(listWithAds)
+    }
+
+    private fun buildListWithAds(movies: List<MovieUiModel>): List<MovieListItem> {
+        val result = mutableListOf<MovieListItem>()
+        movies.forEachIndexed { index, movie ->
+            result.add(MovieListItem.MovieItem(movie))
+            if ((index + 1) % AD_INTERVAL == 0) {
+                result.add(MovieListItem.AdItem(AdUiModel()))
+            }
+        }
+        return result
+    }
+
+    companion object {
+        private const val AD_INTERVAL = 3
     }
 }
